requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
mlflow>=2.8.0
plotly>=5.17.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
supabase>=2.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
